Set lodestar-params as const
