+++ packages/params/setPreset.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-require-imports
+module.exports = require("./lib/setPreset");
