+++ packages/params/src/presetStatus.ts
@@ -1,0 +1,3 @@
+export const presetStatus = {
+  frozen: false,
+};
